+++ src/minecraft/micdoodle8/mods/galacticraft/moon/wgen/GCMoonGenCaves.java
@@ -17,8 +17,8 @@ public class GCMoonGenCaves extends GCCoreMapGenBaseMeta
     
     protected void generateCaveNode(long par1, int par3, int par4, int[] blockIdArray, int[] metaArray, double par6, double par8, double par10, float par12, float par13, float par14, int par15, int par16, double par17)
     {
-        double d4 = (double)(par3 * 16 + 8);
-        double d5 = (double)(par4 * 16 + 8);
+        double d4 = (par3 * 16 + 8);
+        double d5 = (par4 * 16 + 8);
         float f3 = 0.0F;
         float f4 = 0.0F;
         Random random = new Random(par1);
@@ -41,13 +41,13 @@ public class GCMoonGenCaves extends GCCoreMapGenBaseMeta
 
         for (boolean flag1 = random.nextInt(6) == 0; par15 < par16; ++par15)
         {
-            double d6 = 1.5D + (double)(MathHelper.sin((float)par15 * (float)Math.PI / (float)par16) * par12 * 1.0F);
+            double d6 = 1.5D + (MathHelper.sin(par15 * (float)Math.PI / par16) * par12 * 1.0F);
             double d7 = d6 * par17;
             float f5 = MathHelper.cos(par14);
             float f6 = MathHelper.sin(par14);
-            par6 += (double)(MathHelper.cos(par13) * f5);
-            par8 += (double)f6;
-            par10 += (double)(MathHelper.sin(par13) * f5);
+            par6 += (MathHelper.cos(par13) * f5);
+            par8 += f6;
+            par10 += (MathHelper.sin(par13) * f5);
 
             if (flag1)
             {
@@ -67,8 +67,8 @@ public class GCMoonGenCaves extends GCCoreMapGenBaseMeta
 
             if (!flag && par15 == k1 && par12 > 1.0F && par16 > 0)
             {
-                this.generateCaveNode(random.nextLong(), par3, par4, blockIdArray, metaArray, par6, par8, par10, random.nextFloat() * 0.5F + 0.5F, par13 - ((float)Math.PI / 2F), par14 / 3.0F, par15, par16, 1.0D);
-                this.generateCaveNode(random.nextLong(), par3, par4, blockIdArray, metaArray, par6, par8, par10, random.nextFloat() * 0.5F + 0.5F, par13 + ((float)Math.PI / 2F), par14 / 3.0F, par15, par16, 1.0D);
+                this.generateCaveNode(random.nextLong(), par3, par4, blockIdArray, metaArray, par6, par8, par10, random.nextFloat() * 0.5F + 0.5F, par13 - (float)Math.PI / 2F, par14 / 3.0F, par15, par16, 1.0D);
+                this.generateCaveNode(random.nextLong(), par3, par4, blockIdArray, metaArray, par6, par8, par10, random.nextFloat() * 0.5F + 0.5F, par13 + (float)Math.PI / 2F, par14 / 3.0F, par15, par16, 1.0D);
                 return;
             }
 
@@ -76,8 +76,8 @@ public class GCMoonGenCaves extends GCCoreMapGenBaseMeta
             {
                 double d8 = par6 - d4;
                 double d9 = par10 - d5;
-                double d10 = (double)(par16 - par15);
-                double d11 = (double)(par12 + 2.0F + 16.0F);
+                double d10 = (par16 - par15);
+                double d11 = (par12 + 2.0F + 16.0F);
 
                 if (d8 * d8 + d9 * d9 - d10 * d10 > d11 * d11)
                 {
@@ -155,11 +155,11 @@ public class GCMoonGenCaves extends GCCoreMapGenBaseMeta
                     {
                         for (j3 = l1; j3 < i2; ++j3)
                         {
-                            double d12 = ((double)(j3 + par3 * 16) + 0.5D - par6) / d6;
+                            double d12 = ((j3 + par3 * 16) + 0.5D - par6) / d6;
 
                             for (k3 = l2; k3 < i3; ++k3)
                             {
-                                double d13 = ((double)(k3 + par4 * 16) + 0.5D - par10) / d6;
+                                double d13 = ((k3 + par4 * 16) + 0.5D - par10) / d6;
                                 int j4 = (j3 * 16 + k3) * 128 + k2;
                                 boolean flag3 = false;
 
@@ -167,7 +167,7 @@ public class GCMoonGenCaves extends GCCoreMapGenBaseMeta
                                 {
                                     for (int k4 = k2 - 1; k4 >= j2; --k4)
                                     {
-                                        double d14 = ((double)k4 + 0.5D - par8) / d7;
+                                        double d14 = (k4 + 0.5D - par8) / d7;
 
                                         if (d14 > -0.7D && d12 * d12 + d14 * d14 + d13 * d13 < 1.0D)
                                         {
@@ -208,9 +208,9 @@ public class GCMoonGenCaves extends GCCoreMapGenBaseMeta
 
         for (int j1 = 0; j1 < i1; ++j1)
         {
-            double d0 = (double)(par2 * 16 + this.rand.nextInt(16));
-            double d1 = (double)this.rand.nextInt(this.rand.nextInt(120) + 8);
-            double d2 = (double)(par3 * 16 + this.rand.nextInt(16));
+            double d0 = (par2 * 16 + this.rand.nextInt(16));
+            double d1 = this.rand.nextInt(this.rand.nextInt(120) + 8);
+            double d2 = (par3 * 16 + this.rand.nextInt(16));
             int k1 = 1;
 
             if (this.rand.nextInt(4) == 0)
